Restrict filenames to defined alphabets under Unicode print ranges
By shifting the burden to transliterate paths and filenames to the sender
is in O(1), not in O(N) for N depending on the number of mirrors which
would do this needlessly.

The Spatium is accepted despite being a form of space due its
widespread usage in filenames.

In case the user wants to enforce normalization, like everything under
Unicode NFC, this allows to set a Form. It's not always optional, because
some filesystems have bugs with some runes, others offer user-defined
normalizations (ZFS) and can be set to reject anything else, and others
(Apple) just stick with NFD+exceptions despite NFC+exceptions being where
most implementations converge on.