synchronize the LICENSE file(s)
Although this software is not created in a jurisdiction that does not
recognize patents on software, and does not require any or the
"copyright (c)…" line to retain copyright and/or any and all rights,
according to my lawyer I should link anything with any or the LICENSE file:

Unlike in other, unknown to us, jurisdictions, files or binaries which
have not been explicitly licensed may be read without using or storing,
but not used or repurposed in any other way. In other words, it is best
to license everything in a unambiguous way.