log a substitute for a version string
Customers have no way to tell which version of plugin 'upload' they use with
Caddy, as Caddy does not display any date, version, git commit ID, or anything
else to that effect.

Unfortunately I cannot add any scripts to upstream's build server, and the
IDs introduced here do not cover all source code files, but it's a start.

This will enable me to tell if an issue is a duplicate, or if the report
indeed indicates a regression.