use a not-predictable sequence for the random filename suffixes
Originally the sequence has had no guarantee of being unpredictable.
Which I assumed is okay for the home use that Caddy is designed for,
given its inherent design decisions. And then, I used this in integration
tests to check if a file exists.

The latter is not needed anymore, and the former obsolete: Suffixes
cannot be predicted. The overhead for the syscalls is neglible in home
use, and on large production systems you most likely don't want users
to guess filenames anyway.

This can be further optimized by reading a page of random data at a time
and using that as cache for randomness.