CI: run benchmarks on-demand, always try to compile
Unfortunately I don't have a dedicated machine idle to run benchmarks on
at any time. And especially upload speed (which is measured) depends
on machine load and number of other processes running, due to not leveraging
Linux' facilities optimally.

And then, despite lint/vet/gofmt throwing an error I am still interested in the
build artifact, or whether the plugin compiles with the latest version of Caddy
at all. In the past 0.8 → 0.9 of Caddy introduced breaking changes, and I'd like
to be on top of any future ones asap.