protofile: use "default" permission bits on files and directories
Previously more restrictive "permission bits" have been used that could
only be limited, but not expanded to the bitmasks supremum. That is,
there's no "umask" to turn the previous 0750 into a 0777.

This restores the "default behaviour" on a Linux with a commonly used
filesystem and a "usual" access control configuration. On a deviating
configuration (such as the one I use on some servers of mine) those
didn't have any effect anyway.

No tests because Golang's defaults.

closes #31