utilize FADV_{WILLNEED,SEQUENTIAL} for better performance
FADV_SEQUENTIAL indicates that the file will be visited in-order,
and FADV_WILLNEED is supposed to uphold caching.

These settings are esoteric and have little to no (or actually no)
effect on some systems, but using them has no downside save for a
few syscalls at the advantage of boosting performance on the few
systems that support them. For example, anecdotally, on my servers.

Please note that the combination of DONTNEED and NOREUSE decreased
write performance for me.

FADV_SEQUENTIAL in turn didn't increase write speed on my systems,
but did reportedly so for others.