don't write file versions to logs
Utilizing git's "ident" facility was needed to assign IDs to source files
that contributed to a binary, to tell whether the latest version of this
plugin was used on bug reports.

Since Caddy's switch to using Go modules this is obsolete.
Go 1.12's own runtime/debug#ReadBuildInfo will give a list and versions.

See also:
  * https://golang.org/pkg/runtime/debug/#BuildInfo