CI: build utilizing Go modules
Essentially that's setting environment variable GO111MODULE=on because
the folder structure is still for GOPATH, but retained to segregate our
sources and their's (for later caching).

Golang fetches any missing dependencies automatically on any `go build`
or `go test`, hence all `go get…` have been replaced.

Unlike with final releases, the versions in `go.mod` files are off and
we need the `replace … => /…/…` lines to point the compiler to where
source code resides that is to be used.