CI: use Google's Ubuntu image for integration tests
For one, Docker's image hub does not need to be queried anymore, which
eliminates the vulnerability of relying on a set of third parties
(internet connections, Docker, the hub's infrastructure provider).

Further, absolutely rules out a change of mine only found in one of
my images were needed for this plugin to work.

A new downside is Google's curated image with Ubuntu comes with rather
old libraries, and could be lacking tools or features in tools,
which could become relevant for integration testing.