Show how to upload to cloud storage buckets
The Go CDK v0.23 is able to skip sidecar files with "?metadata=skip" which
is a requisite for an image hoster. Since that is in and released,
release v5 as well. Any user replaying the documentation will arrive
at a installation working as intended.

For examples on how to use signed URLs instead, see:
 * https://cloud.google.com/blog/products/storage-data-transfer/uploading-images-directly-to-cloud-storage-by-using-signed-url
 * https://docs.min.io/docs/upload-files-from-browser-using-pre-signed-urls.html